+++ mpvpaper/BarWidget.qml
@@ -2,42 +2,58 @@ import QtQuick
 import QtQuick.Layouts
 import Quickshell
 import qs.Commons
+import qs.Modules.Bar.Extras
 import qs.Services.UI
 import qs.Widgets
 
-NIconButton {
+Item {
     id: root
 
     property var pluginApi: null
     property ShellScreen screen
-    
+
     readonly property bool active: 
         pluginApi.pluginSettings.active || 
         false
 
-    icon: "wallpaper-selector"
+    readonly property bool isPlaying:
+        pluginApi.pluginSettings.isPlaying ||
+        false
 
-    onClicked: {
-        pluginApi?.openPanel(root.screen, root);
-    }
+    readonly property bool isMuted:
+        pluginApi.pluginSettings.isMuted ||
+        false
 
-    onRightClicked: {
-        PanelService.showContextMenu(contextMenu, root, screen);
-    }
-    
+    implicitWidth: pill.width
+    implicitHeight: pill.height
+
     NPopupContextMenu {
         id: contextMenu
 
         model: [
             {
-                "label": pluginApi?.tr("barWidget.contextMenu.panel") || "Panel",
+                "label": root.pluginApi?.tr("barWidget.contextMenu.panel") || "Panel",
                 "action": "panel",
                 "icon": "rectangle"
             },
             {
-                "label": pluginApi?.tr("barWidget.contextMenu.toggle") || "Toggle",
+                "label": root.pluginApi?.tr("barWidget.contextMenu.toggle") || "Toggle",
                 "action": "toggle",
                 "icon": "power"
+            },
+            {
+                "label": root.isPlaying ? 
+                    root.pluginApi?.tr("barWidget.contextMenu.pause") || "Pause" : 
+                    root.pluginApi?.tr("barWidget.contextMenu.play") || "Play",
+                "action": root.isPlaying ? "pause" : "play",
+                "icon": root.isPlaying ? "media-pause" : "media-play"
+            },
+            {
+                "label": root.isMuted ? 
+                    root.pluginApi?.tr("barWidget.contextMenu.unmute") || "Unmute" : 
+                    root.pluginApi?.tr("barWidget.contextMenu.mute") || "Mute",
+                "action": root.isMuted ? "unmute" : "mute",
+                "icon": root.isMuted ? "volume-high" : "volume-mute"
             }
         ]
 
@@ -45,12 +61,50 @@ NIconButton {
             contextMenu.close();
             PanelService.closeContextMenu(root.screen);
 
-            if(action === "panel") {
-                pluginApi?.openPanel(root.screen, root);
-            } else if (action === "toggle") {
-                pluginApi.pluginSettings.active = !root.active;
-                pluginApi.saveSettings();
+            switch(action) {
+                case "panel":
+                    root.pluginApi?.openPanel(root.screen, root);
+                    break;
+                case "toggle":
+                    root.pluginApi.pluginSettings.active = !root.active;
+                    root.pluginApi.saveSettings();
+                    break;
+                case "play":
+                    root.pluginApi.pluginSettings.isPlaying = true;
+                    root.pluginApi.saveSettings();
+                    break;
+                case "pause":
+                    root.pluginApi.pluginSettings.isPlaying = false;
+                    root.pluginApi.saveSettings();
+                    break;
+                case "mute":
+                    root.pluginApi.pluginSettings.isMuted = true;
+                    root.pluginApi.saveSettings();
+                    break;
+                case "unmute":
+                    root.pluginApi.pluginSettings.isMuted = false;
+                    root.pluginApi.saveSettings();
+                    break;
+                default:
+                    Logger.e("mpvpaper", "Error, action not found:", action);
             }
+        }
+    }
+
+    BarPill {
+        id: pill
+
+        screen: root.screen
+        tooltipText: "Open mpvpaper manager"
+
+        icon: "wallpaper-selector"
+
+        onClicked: {
+            pluginApi?.openPanel(root.screen, root);
+        }
+
+        onRightClicked: {
+            PanelService.showContextMenu(contextMenu, root, screen);
         }
     }
 }